requirements.txt
# Core Dependencies
aiohttp==3.9.1
anyio==4.4.0
attrs==23.2.0

# Web Framework
fastapi==0.100.0
uvicorn==0.22.0
langfuse==2.59.7
python-multipart==0.0.6
streamlit==1.30.0
# Threading/Concurrency
tqdm==4.66.1

# Date handling
python-dateutil==2.8.2

# For using OpenAI functions
jsonschema==4.20.0
# Database
sqlalchemy==2.0.31
psycopg2
duckdb==0.8.1

# ML & Analysis
pandas==2.0.3
numpy==1.23.5
# Use a prebuilt SciPy wheel
scipy==1.10.1
scikit-learn==1.3.0
spacy==3.7.4
nltk

# AI & LangChain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
openai==1.10.0
groq==0.15.0
instructor==1.2.6
chromadb==0.4.22
tiktoken==0.5.2

# Visualization
plotly==5.18.0

# Utils
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.10.0
beautifulsoup4==4.12.3
debugpy==1.8.0
pyjwt==2.7.0
psycopg2-binary==2.9.9  # Use a newer version


# Testing
pytest==7.4.0
httpx==0.27.0